mdm: Refactor switching instance in MultiDieManager
Moved the responsibility to switch an instance to a helper class which handles the whole process of finding the assigned die, inheriting the placement info, net info, properties, and group info from the original instance. Also, a method added to create a die according to the existing technologies in the database except the first two. These changes allow better organization of code and improved flexibility for managing Multi-die scenarios.